requirements.txt
asgiref==3.6.0
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
distro==1.7.0
Django==4.1.7
django-debug-toolbar==4.2.0
django-filter==23.2
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.4
fasteners==0.14.1
flake8==6.1.0
flake8-docstrings==1.7.0
future==0.18.2
httplib2==0.20.2
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.11.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.1.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.1
pytz==2022.1
pyxdg==0.27
referencing==0.30.2
requests==2.25.1
rpds-py==0.9.2
SecretStorage==3.3.1
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.3
